requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
websockets>=11.0.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
